Remove `include/grpcpp/impl/codegen/server_callback.h`
