+++ src/compiler/cpp_generator.cc
@@ -147,7 +147,7 @@ std::string GetHeaderIncludes(grpc_generator::File* file,
         "grpcpp/support/method_handler.h",
         "grpcpp/impl/codegen/proto_utils.h",
         "grpcpp/impl/codegen/rpc_method.h",
-        "grpcpp/impl/codegen/server_callback.h",
+        "grpcpp/support/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
         "grpcpp/impl/codegen/server_context.h",
         "grpcpp/impl/codegen/service_type.h",
@@ -1656,7 +1656,7 @@ std::string GetSourceIncludes(grpc_generator::File* file,
         "grpcpp/support/message_allocator.h",
         "grpcpp/support/method_handler.h",
         "grpcpp/impl/codegen/rpc_service_method.h",
-        "grpcpp/impl/codegen/server_callback.h",
+        "grpcpp/support/server_callback.h",
         "grpcpp/impl/codegen/server_callback_handlers.h",
         "grpcpp/impl/codegen/server_context.h",
         "grpcpp/impl/codegen/service_type.h",
